+++ src/app/components/cluster-dns.js
@@ -10,7 +10,7 @@
 
 import Component from '@ember/component';
 import I18n from 'onedata-gui-common/mixins/components/i18n';
-import EmberObject, { computed, get, set } from '@ember/object';
+import EmberObject, { computed, observer, get, set } from '@ember/object';
 import { reads } from '@ember/object/computed';
 import safeExec from 'onedata-gui-common/utils/safe-method-execution';
 import { inject as service } from '@ember/service';
@@ -42,6 +42,7 @@ export default Component.extend(
     dnsManager: service(),
     guiUtils: service(),
     globalNotify: service(),
+    i18n: service(),
 
     i18nPrefix: 'components.clusterDns',
 
@@ -69,6 +70,14 @@ export default Component.extend(
      * Invoked withoud params when settings changes (check is not current anymore)
      */
     settingsChanged: notImplementedIgnore,
+
+    /**
+     * @virtual optional
+     * Invoked with isIpDomain value when isIpDomain changed.
+     * @type {function}
+     * @param {boolean} isIpChanged
+     */
+    isIpDomainChanged: notImplementedIgnore,
 
     /**
      * @type {EmberObject}
@@ -111,6 +120,18 @@ export default Component.extend(
      * @type {boolean}
      */
     subdomainDelegationPrev: undefined,
+
+    /**
+     * Contains old value of dnsServers to preserve dns servers input state
+     * after change to autodetect mode.
+     * @type {Array<string>}
+     */
+    dnsServersPrev: Object.freeze([]),
+
+    /**
+     * @type {string}
+     */
+    dnsCheckMode: 'autodetect',
 
     /**
      * @type {Ember.ComputedProperty<string>}
@@ -257,9 +278,14 @@ export default Component.extend(
      * True if the list in tokenized input of DNS servers is valid.
      * @type {Ember.ComputedProperty<boolean>}
      */
-    dnsServersInputValid: computed('dnsServers.[]', function dnsServersInputValid() {
-      return !_.isEmpty(this.get('dnsServers'));
-    }),
+    dnsServersInputValid: computed(
+      'dnsServers.[]',
+      'dnsCheckMode',
+      function dnsServersInputValid() {
+        return !_.isEmpty(this.get('dnsServers')) ||
+          this.get('dnsCheckMode') === 'autodetect';
+      }
+    ),
 
     /**
      * Each object is a modified clone of Onepanel.DnsCheckResult with type added
@@ -317,6 +343,27 @@ export default Component.extend(
       return checkResultItems.every(i => get(i, 'summary') === 'ok');
     }),
 
+    /**
+     * @type {Ember.ComputedProperty<Array<Object>>}
+     */
+    dnsCheckModes: computed(function dnsCheckModes() {
+      return [{
+        label: this.t('dnsCheckAutodetect'),
+        value: 'autodetect',
+      }, {
+        label: this.t('dnsCheckManual'),
+        value: 'manual',
+      }];
+    }),
+
+    isIpDomainObserver: observer('isIpDomain', function isIpDomainObserver() {
+      const {
+        isIpDomainChanged,
+        isIpDomain,
+      } = this.getProperties('isIpDomainChanged', 'isIpDomain');
+      isIpDomainChanged(isIpDomain);
+    }),
+
     init() {
       this._super(...arguments);
       if (!this.get('formValues')) {
@@ -324,7 +371,11 @@ export default Component.extend(
           letsEncrypt: true,
         }));
       }
-      this.updateDnsCheckConfigurationProxy();
+      this.updateDnsCheckConfigurationProxy()
+        .then(({ dnsServers }) => safeExec(this, () => {
+          this.set('dnsCheckMode', get(dnsServers, 'length') ? 'manual' :
+            'autodetect');
+        }));
       this.updateDomainProxy();
       if (this.get('getDnsCheckProxyOnStart')) {
         (
@@ -344,6 +395,8 @@ export default Component.extend(
       } else {
         this.updateProviderProxy();
       }
+      // enable isIpDomainObserver 
+      this.get('isIpDomain');
     },
 
     willDestroyElement() {
@@ -535,6 +588,18 @@ export default Component.extend(
             dnsServers: _.without(dnsServersUpdate, newIpAddress),
             newIpAddress,
           });
+        }
+      },
+      dnsCheckModeChanged(value) {
+        this.set('dnsCheckMode', value);
+        if (value === 'autodetect') {
+          this.set('dnsServersPrev', this.get('dnsServers'));
+          this.send('dnsServersChanged', []);
+        } else {
+          const dnsServersPrev = this.get('dnsServersPrev');
+          if (get(dnsServersPrev, 'length')) {
+            this.send('dnsServersChanged', this.get('dnsServersPrev'));
+          }
         }
       },
     },